Move the base keymap and multibuffer hint to onboarding crate
