+++ crates/onboarding/Cargo.toml
@@ -27,6 +27,7 @@ fs.workspace = true
 fuzzy.workspace = true
 git.workspace = true
 gpui.workspace = true
+install_cli.workspace = true
 itertools.workspace = true
 language.workspace = true
 language_model.workspace = true
@@ -37,6 +38,7 @@ project.workspace = true
 schemars.workspace = true
 serde.workspace = true
 settings.workspace = true
+telemetry.workspace = true
 theme.workspace = true
 ui.workspace = true
 util.workspace = true
