+++ crates/onboarding/src/base_keymap_picker.rs
@@ -12,7 +12,7 @@ use util::ResultExt;
 use workspace::{ModalView, Workspace, ui::HighlightedLabel};
 
 actions!(
-    welcome,
+    zed,
     [
         /// Toggles the base keymap selector modal.
         ToggleBaseKeymapSelector
